+++ .idea/workspace.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ChangeListManager">
+    <list default="true" id="652357e5-1e1e-4eaf-ba2f-f5e1b8b5a672" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/.idea/.gitignore" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/Fitness/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/Fitness/asgi.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/Fitness/settings.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/Fitness/urls.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/Fitness/wsgi.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/admin.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/apps.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/migrations/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/models.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/serializers.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/tests.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/FitnessApp/views.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/db.sqlite3" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/Fitness/manage.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/backend/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/Fitness.sqlite3" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp/asgi.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp/settings.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp/urls.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp/wsgi.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/admin.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/apps.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/migrations/0001_initial.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/migrations/0002_exercise_training_user_statistics.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/migrations/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/models.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/serializers.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/tests.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/urls.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/FitnessApp1/views.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/db.sqlite3" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/FitnessApp/manage.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/admin.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/apps.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/migrations/__init__.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/models.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/tests.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/frontend/views.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Fitness/main.py" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Python Script" />
+      </list>
+    </option>
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectId" id="2OjXNLdOO2nE1XMAbW4Srjp5efF" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "last_opened_file_path": "C:/Users/Salko/Desktop/учеба тпу/Fitness",
+    "settings.editor.selected.configurable": "com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable"
+  }
+}]]></component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\Salko\Desktop\учеба тпу\Fitness-website\backend" />
+    </key>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="652357e5-1e1e-4eaf-ba2f-f5e1b8b5a672" name="Changes" comment="" />
+      <created>1682074923811</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1682074923811</updated>
+    </task>
+    <servers />
+  </component>
+</project>
