--- Fitness/FitnessApp/FitnessApp/urls.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from django.urls import path, include
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('', include('FitnessApp1.urls')),
-]
