--- Fitness/FitnessApp/FitnessApp1/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import User
-
-admin.site.register(User)
