--- Fitness/FitnessApp/FitnessApp1/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class Fitnessapp1Config(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'FitnessApp1'
