+++ backend/Fitness/Fitness/wsgi.py
@@ -1,0 +1,16 @@
+"""
+WSGI config for Fitness project.
+
+It exposes the WSGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/4.2/howto/deployment/wsgi/
+"""
+
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Fitness.settings')
+
+application = get_wsgi_application()
