+++ backend/Fitness/FitnessApp/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FitnessappConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'FitnessApp'
